+++ service/src/main/kotlin/fi/espoo/evaka/Audit.kt
@@ -58,7 +58,6 @@ enum class Audit(
     ApplicationVerify,
     ApplicationsReportRead,
     AssistanceActionOptionsRead,
-    AssistanceBasisOptionsRead,
     AssistanceFactorCreate,
     AssistanceFactorUpdate,
     AssistanceFactorDelete,
@@ -93,7 +92,6 @@ enum class Audit(
     CalendarEventUpdate,
     CalendarEventTimeCreate,
     CalendarEventTimeDelete,
-    CalendarEventTimeRead,
     CalendarEventChildTimesCancellation,
     CalendarEventTimeReservationCreate,
     CalendarEventTimeReservationDelete,
@@ -104,9 +102,6 @@ enum class Audit(
     ChildAssistanceActionCreate,
     ChildAssistanceActionDelete,
     ChildAssistanceActionUpdate,
-    ChildAssistanceNeedCreate,
-    ChildAssistanceNeedDelete,
-    ChildAssistanceNeedUpdate,
     ChildAssistanceNeedDecisionAnnul,
     ChildAssistanceNeedDecisionCreate,
     ChildAssistanceNeedDecisionDelete,
@@ -140,7 +135,6 @@ enum class Audit(
     ChildAssistanceNeedPreschoolDecisionRevertToUnsent,
     ChildAssistanceNeedPreschoolDecisionDecide,
     ChildAssistanceNeedPreschoolDecisionOpened,
-    ChildAssistanceNeedPreschoolDecisionUpdateDecisionMaker,
     ChildAssistanceNeedVoucherCoefficientCreate,
     ChildAssistanceNeedVoucherCoefficientRead,
     ChildAssistanceNeedVoucherCoefficientUpdate,
@@ -148,7 +142,6 @@ enum class Audit(
     ChildAttendanceChildrenRead,
     ChildAttendanceReportRead,
     ChildAttendanceStatusesRead,
-    ChildAttendancesUpsert,
     ChildAttendancesArrivalCreate,
     ChildAttendancesDepartureRead,
     ChildAttendancesDepartureCreate,
@@ -218,7 +211,6 @@ enum class Audit(
     CitizenEmailVerificationStatusRead,
     CitizenFeeDecisionDownloadPdf,
     CitizenNotificationSettingsRead,
-    CitizenNotificationSettingsUpdate,
     CitizenLogin(securityEvent = true, securityLevel = "high"),
     CitizenCredentialsUpdate(securityEvent = true, securityLevel = "high"),
     CitizenCredentialsUpdateAttempt(securityEvent = true, securityLevel = "high"),
@@ -240,7 +232,6 @@ enum class Audit(
     DaycareGroupPlacementCreate,
     DaycareGroupPlacementDelete,
     DaycareGroupPlacementTransfer,
-    DaycareBackupCareRead,
     DecisionAccept,
     DecisionDownloadPdf,
     DecisionDraftRead,
@@ -285,7 +276,6 @@ enum class Audit(
     FeeDecisionHeadOfFamilyRead,
     FeeDecisionHeadOfFamilyCreateRetroactive,
     FeeDecisionIgnore,
-    FeeDecisionLiableCitizenRead,
     FeeDecisionMarkSent,
     FeeDecisionPdfRead,
     FeeDecisionRead,
@@ -359,7 +349,6 @@ enum class Audit(
     InvoicesSearch,
     InvoicesSend,
     InvoicesSendByDate,
-    InvoicesUpdate,
     ManualDuplicationReportRead,
     MealReportRead,
     MessagingMyAccountsRead,
@@ -439,7 +428,6 @@ enum class Audit(
     PersonDelete(securityEvent = true, securityLevel = "high"),
     PersonDependantRead(securityEvent = true, securityLevel = "high"),
     PersonGuardianRead(securityEvent = true, securityLevel = "high"),
-    PersonBlockedGuardiansRead(securityEvent = true, securityLevel = "high"),
     PersonDetailsRead(securityEvent = true, securityLevel = "high"),
     PersonDetailsSearch,
     PersonDuplicate,
@@ -463,7 +451,6 @@ enum class Audit(
     PlacementPlanCreate,
     PlacementPlanRespond,
     PlacementPlanDraftRead,
-    PlacementPlanSearch,
     PlacementProposalCreate,
     PlacementProposalAccept,
     PlacementSearch,
@@ -482,7 +469,6 @@ enum class Audit(
     PreschoolTermCreate,
     PreschoolTermUpdate,
     PreschoolTermDelete,
-    PreschoolTermRead,
     PresenceReportRead,
     PushSettingsRead,
     PushSettingsSet,
@@ -505,8 +491,6 @@ enum class Audit(
     StaffAttendanceDepartureExternalCreate,
     StaffAttendanceRead,
     StaffAttendanceUpdate,
-    StaffAttendanceDelete,
-    StaffAttendanceExternalDelete,
     StaffAttendanceExternalUpdate,
     StaffOccupancyCoefficientRead,
     StaffOccupancyCoefficientUpsert,
