--- service/src/main/kotlin/fi/espoo/evaka/invoicing/controller/ControllerUtil.kt
@@ -1,15 +1,0 @@
-// SPDX-FileCopyrightText: 2017-2020 City of Espoo
-//
-// SPDX-License-Identifier: LGPL-2.1-or-later
-
-package fi.espoo.evaka.invoicing.controller
-
-import fi.espoo.evaka.shared.domain.BadRequest
-import java.util.UUID
-
-fun parseUUID(uuid: String): UUID =
-    try {
-        UUID.fromString(uuid)
-    } catch (e: IllegalArgumentException) {
-        throw BadRequest("Given ID ($uuid) is not a valid UUID")
-    }
