+++ service/src/main/kotlin/fi/espoo/evaka/serviceneed/ServiceNeedQueries.kt
@@ -385,26 +385,6 @@ ORDER BY child_id, valid_during
         .toList<ChildServiceNeedInfo>()
 }
 
-fun Database.Read.getServiceNeedOptionFees(from: LocalDate): List<ServiceNeedOptionFee> {
-    return createQuery {
-            sql(
-                """
-SELECT
-    service_need_option_id,
-    validity,
-    base_fee,
-    sibling_discount_2,
-    sibling_fee_2,
-    sibling_discount_2_plus,
-    sibling_fee_2_plus
-FROM service_need_option_fee
-WHERE validity && daterange(${bind(from)}, null, '[]')
-"""
-            )
-        }
-        .toList<ServiceNeedOptionFee>()
-}
-
 fun Database.Read.getServiceNeedOptionFees(): List<ServiceNeedOptionFee> {
     return createQuery {
             sql(
