--- service/src/main/kotlin/fi/espoo/evaka/shared/controllers/Wrapper.kt
@@ -1,7 +1,0 @@
-// SPDX-FileCopyrightText: 2017-2022 City of Espoo
-//
-// SPDX-License-Identifier: LGPL-2.1-or-later
-
-package fi.espoo.evaka.shared.controllers
-
-data class Wrapper<T>(val data: T)
