+++ service/src/main/kotlin/fi/espoo/evaka/shared/db/Database.kt
@@ -458,74 +458,6 @@ class Database(private val jdbi: Jdbi, private val tracer: Tracer) {
         }
     }
 
-    class LegacyPreparedBatch
-    internal constructor(private val raw: org.jdbi.v3.core.statement.PreparedBatch) {
-
-        fun add(): LegacyPreparedBatch {
-            raw.add()
-            return this
-        }
-
-        fun execute(): IntArray = raw.execute()
-
-        fun executeAndReturn(): UpdateResult = UpdateResult(raw.executePreparedBatch())
-
-        inline fun <reified T> bind(name: String, value: T): LegacyPreparedBatch =
-            bindByType(name, value, createQualifiedType(*defaultQualifiers<T>()))
-
-        inline fun <reified T> registerColumnMapper(mapper: ColumnMapper<T>): LegacyPreparedBatch =
-            registerColumnMapper(createQualifiedType(), mapper)
-
-        fun <T> registerColumnMapper(
-            type: QualifiedType<T>,
-            mapper: ColumnMapper<T>,
-        ): LegacyPreparedBatch {
-            raw.registerColumnMapper(type, mapper)
-            return this
-        }
-
-        fun addBinding(binding: Binding<*>): LegacyPreparedBatch {
-            raw.bindByType(binding.name, binding.value, binding.type)
-            return this
-        }
-
-        fun addBindings(bindings: Iterable<Binding<*>>): LegacyPreparedBatch {
-            for (binding in bindings) {
-                raw.bindByType(binding.name, binding.value, binding.type)
-            }
-            return this
-        }
-
-        fun addBindings(bindings: Sequence<Binding<*>>): LegacyPreparedBatch {
-            for (binding in bindings) {
-                raw.bindByType(binding.name, binding.value, binding.type)
-            }
-            return this
-        }
-
-        fun bindJson(name: String, value: Any): LegacyPreparedBatch =
-            bindByType(
-                name,
-                value,
-                QualifiedType.of(value.javaClass).withAnnotationClasses(listOf(Json::class.java)),
-            )
-
-        fun <T> bindByType(name: String, value: T, type: QualifiedType<T>): LegacyPreparedBatch {
-            raw.bindByType(name, value, type)
-            return this
-        }
-
-        fun bindKotlin(value: Any): LegacyPreparedBatch {
-            raw.bindKotlin(value)
-            return this
-        }
-
-        fun bindKotlin(name: String, value: Any): LegacyPreparedBatch {
-            raw.bindKotlin(name, value)
-            return this
-        }
-    }
-
     class PreparedBatch<R>
     internal constructor(
         private val raw: org.jdbi.v3.core.statement.PreparedBatch,
